pyproject.toml
[tool.poetry]
name = "krunner-keepassxc"
version = "1.0.0"
description = ""
authors = ["naglfar <stefan@mtztech.info>"]

[tool.poetry.dependencies]
python = "^3.6,<3.9"
dbus-python = "^1.2.16"
cryptography = "^2.9.2"

[tool.poetry.dev-dependencies]
vext = "^0.7.3"
"vext.gi" = "^0.7.0"
pytest = "^5.2"
pex = "^2.1.11"
shiv = "^0.1.2"
radon = "^4.1.0"
taskipy = "^1.2.1"
mypy = "^0.770"

[tool.poetry.scripts]
runner = "krunner_keepassxc.__main__:main"
cli = "krunner_keepassxc.__main__:cli"

[tool.taskipy.tasks]
test = "python -m unittest tests/test_*.py"
mypy = "mypy --ignore-missing-imports --check-untyped-defs krunner_keepassxc tests"
radon = "radon cc krunner_keepassxc"
pex = """pex \
		--disable-cache \
		--inherit-path \
		-r pex_requirements.txt \
		-e krunner_keepassxc \
		-o dist/krunner_keepassxc.pex \
		."""
shiv = """shiv \
		-p "/usr/bin/env python3" \
		-e krunner_keepassxc.__main__:main \
		--output-file dist/krunner_keepassxc.pyz \
		."""
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

